feat(app): added insert user page
