+++ src/sections/user/view/user-view.tsx
@@ -1,216 +1,34 @@
-import { useState, useCallback, useEffect } from 'react';
-
-import Box from '@mui/material/Box';
-import Card from '@mui/material/Card';
-import Table from '@mui/material/Table';
-import Button from '@mui/material/Button';
-import TableBody from '@mui/material/TableBody';
-import Typography from '@mui/material/Typography';
-import TableContainer from '@mui/material/TableContainer';
-import TablePagination from '@mui/material/TablePagination';
+import React, { useState } from 'react';
 
 import { _users } from 'src/_mock';
 import { DashboardContent } from 'src/layouts/dashboard';
-
-import { Iconify } from 'src/components/iconify';
-import { Scrollbar } from 'src/components/scrollbar';
-
-import { TableNoData } from '../table-no-data';
-import { UserTableRow } from '../user-table-row';
-import { UserTableHead } from '../user-table-head';
-import { TableEmptyRows } from '../table-empty-rows';
-import { UserTableToolbar } from '../user-table-toolbar';
-import { emptyRows, applyFilter, getComparator } from '../utils';
-
-import type { UserProps } from '../user-table-row';
-import axios from 'axios';
-import { useToaster } from 'src/components/toast/Toast';
+import { InsertUserView } from './insert-user-view';
+import { ShowUserView } from './show-user-view';
+import { UserPageNumbers } from '../utils';
 
 // ----------------------------------------------------------------------
 
 export function UserView() {
-  const table = useTable();
-  const { showErrorToast, showSuccessToast } = useToaster();
-
+  const [currPage, setCurrPage] = useState<number>(1);
   const [updateSignal, setUpdateSignal] = useState<boolean>(false);
 
-  const [users, setUsers] = useState<UserProps[]>([]);
-  const [filterName, setFilterName] = useState<string>('');
-
-  const dataFiltered: UserProps[] = applyFilter({
-    inputData: users,
-    comparator: getComparator(table.order, table.orderBy),
-    filterInput: filterName,
-  });
-
-  useEffect(() => {
-    const getUsers = async () => {
-      try {
-        const response = await axios.get(`${import.meta.env.VITE_BACKEND_API}${import.meta.env.VITE_API_ENDPOINT_USER}`);
-        setUsers(response.data);
-        console.log(response.data);
-      } catch (error) {
-        showErrorToast(error.message);
+  const renderSubpage: (page: number) => React.JSX.Element = (page: number) => {
+    switch (page) {
+      case UserPageNumbers.INSERT_PAGE_VIEW: { // insert view
+        return <InsertUserView currPage={currPage} changePage={setCurrPage} updateSignal={updateSignal} handleUpdate={() => setUpdateSignal(!updateSignal)} />;
       }
-    };
-
-    getUsers();
-  }, [updateSignal]);
-
-  const notFound = !dataFiltered.length && !!filterName;
+      case UserPageNumbers.SHOW_PAGE_VIEW: { // show view
+        return <ShowUserView currPage={currPage} changePage={setCurrPage} updateSignal={updateSignal} handleUpdate={() => setUpdateSignal(!updateSignal)} />;
+      }
+      default: { // return nothing if unknown page number
+        return <></>;
+      }
+    }
+  }
 
   return (
     <DashboardContent>
-      <Box display="flex" alignItems="center" mb={5}>
-        <Typography variant="h4" flexGrow={1}>
-          Users
-        </Typography>
-        <Button
-          variant="contained"
-          color="inherit"
-          startIcon={<Iconify icon="mingcute:add-line" />}
-        >
-          New user
-        </Button>
-      </Box>
-
-      <Card>
-        <UserTableToolbar
-          numSelected={table.selected.length}
-          filterName={filterName}
-          onFilterName={(event: React.ChangeEvent<HTMLInputElement>) => {
-            setFilterName(event.target.value);
-            table.onResetPage();
-          }}
-        />
-
-        <Scrollbar>
-          <TableContainer sx={{ overflow: 'unset' }}>
-            <Table sx={{ minWidth: 800 }}>
-              <UserTableHead
-                order={table.order}
-                orderBy={table.orderBy}
-                rowCount={_users.length}
-                numSelected={table.selected.length}
-                onSort={table.onSort}
-                onSelectAllRows={(checked) =>
-                  table.onSelectAllRows(
-                    checked,
-                    _users.map((user) => user.id)
-                  )
-                }
-                headLabel={[
-                  { id: 'name', label: 'Name' },
-                  { id: 'email', label: 'E-mail' },
-                  { id: 'phone', label: 'Phone Number' },
-                  { id: 'role', label: 'Role' },
-                  { id: 'action', label: 'Action', align: 'left'},
-                ]}
-              />
-              <TableBody>
-                {dataFiltered
-                  .slice(
-                    table.page * table.rowsPerPage,
-                    table.page * table.rowsPerPage + table.rowsPerPage
-                  )
-                  .map((row) => (
-                    <UserTableRow
-                      key={row.userId}
-                      row={row}
-                      selected={table.selected.includes(row.userId)}
-                      onSelectRow={() => table.onSelectRow(row.userId)}
-                    />
-                  ))}
-
-                <TableEmptyRows
-                  height={68}
-                  emptyRows={emptyRows(table.page, table.rowsPerPage, _users.length)}
-                />
-
-                {notFound && <TableNoData searchQuery={filterName} />}
-              </TableBody>
-            </Table>
-          </TableContainer>
-        </Scrollbar>
-
-        <TablePagination
-          component="div"
-          page={table.page}
-          count={_users.length}
-          rowsPerPage={table.rowsPerPage}
-          onPageChange={table.onChangePage}
-          rowsPerPageOptions={[5, 10, 25]}
-          onRowsPerPageChange={table.onChangeRowsPerPage}
-        />
-      </Card>
+      { renderSubpage(currPage) }
     </DashboardContent>
   );
-}
-
-// ----------------------------------------------------------------------
-
-export function useTable() {
-  const [page, setPage] = useState(0);
-  const [orderBy, setOrderBy] = useState('name');
-  const [rowsPerPage, setRowsPerPage] = useState(5);
-  const [selected, setSelected] = useState<string[]>([]);
-  const [order, setOrder] = useState<'asc' | 'desc'>('asc');
-
-  const onSort = useCallback(
-    (id: string) => {
-      const isAsc = orderBy === id && order === 'asc';
-      setOrder(isAsc ? 'desc' : 'asc');
-      setOrderBy(id);
-    },
-    [order, orderBy]
-  );
-
-  const onSelectAllRows = useCallback((checked: boolean, newSelecteds: string[]) => {
-    if (checked) {
-      setSelected(newSelecteds);
-      return;
-    }
-    setSelected([]);
-  }, []);
-
-  const onSelectRow = useCallback(
-    (inputValue: string) => {
-      const newSelected = selected.includes(inputValue)
-        ? selected.filter((value) => value !== inputValue)
-        : [...selected, inputValue];
-
-      setSelected(newSelected);
-    },
-    [selected]
-  );
-
-  const onResetPage = useCallback(() => {
-    setPage(0);
-  }, []);
-
-  const onChangePage = useCallback((event: unknown, newPage: number) => {
-    setPage(newPage);
-  }, []);
-
-  const onChangeRowsPerPage = useCallback(
-    (event: React.ChangeEvent<HTMLInputElement>) => {
-      setRowsPerPage(parseInt(event.target.value, 10));
-      onResetPage();
-    },
-    [onResetPage]
-  );
-
-  return {
-    page,
-    order,
-    onSort,
-    orderBy,
-    selected,
-    rowsPerPage,
-    onSelectRow,
-    onResetPage,
-    onChangePage,
-    onSelectAllRows,
-    onChangeRowsPerPage,
-  };
 }
