+++ src/vite-env.d.ts
@@ -1,7 +1,7 @@
 /// <reference types="vite/client" />
 
 interface ImportMetaEnv {
-    readonly API_ENDPOINT_USER: string;
+    readonly VITE_API_ENDPOINT_USER: string;
 }
 
 interface ImportMeta {
